Initial commit.
This commit adds the initial version of the project, which includes the following files:

Matrix.cs: A class that contains methods for generating integer matrices, copying matrices, and splitting matrices into equal or nearly equal parts.

MatrixRowSorting.cs: A class that contains a method for sorting the rows of a matrix using multiple threads.

Benchmark.cs: A class that will be used for running the sorting methods in various thread configurations and measuring their execution time. This includes two methods, RunSingleTest and RunMultipleTests, which take input parameters and return the execution time or a list of execution times, respectively.

Project configuration files and .gitignore.